Posts Create view controller part2
Additions

When designing the views in your application take a moment to consider what type of user will be interacting with the view. The level of styling and customization needed for the view will be tied to the intended user. For the general public you will want to be as explicit as possible. For an internal application that users will be trained on you can assume a level of competency. For a view that only you the developer will use, the requirements are simply what you need.

Placing yourself in the role of the intended user will result in better applications.

When creating a record that has one or more parents each of the foreign key properties must be a valid primary key that is already in the database. This is why foreign key properties are created with a dropdown menu instead of an input.